Create Slideractionusing_Selenium
The objective of this program is to demonstrate the use of Selenium in order to demonstrate slider action. For this purpose we will be using jquery. Information gleaned from web as well as online courses. Please feel free to modify for your practice.

The program structure begins by importing some essential libraries and invoking some methods that are vital for this task.