Update and rename searchforproducts to searchforproducts_selenium
The objective of this Python script is to open up a Firefox session ( or whatever browser one chooses) that uses the Selenium WebDriver client library and does some operations after the website is hit and some search items are returned. Information gleaned from online courses as well as specific books on application of Selenium Webdriver as it pertains to Python.